+++ hahomematic/platforms/generic/entity.py
@@ -141,7 +141,7 @@ class GenericEntity[ParameterT: GenericParameterType, InputParameterT: GenericPa
         if self._forced_usage:
             return self._forced_usage
         if self._central.parameter_visibility.parameter_is_hidden(
-            device_type=self._device.model,
+            model=self._device.model,
             channel_no=self.channel_no,
             paramset_key=self._paramset_key,
             parameter=self._parameter,