+++ changelog.md
@@ -1,9 +1,15 @@
+# Version 2024.9.8 (2024-09-21)
+
+- Refactor device/entity to extract channel
+- Replace device_type by model
+- Shorten names
+-
+
 # Version 2024.9.7 (2024-09-15)
 
 - Add bind_collector to all relevant methods with option to disable it
 - Identify bind_collector annotated methods
 - Mark externally accessed services with service_call if bind_collector is not appropriate
-- Refactor device/entity to extract channel
 
 # Version 2024.9.6 (2024-09-13)